server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
attrs==24.3.0
azure-core==1.32.0
azure-cosmos==4.9.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.35.97
botocore==1.35.97
browser-use==0.1.19
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
courlan==1.3.2
dateparser==1.2.0
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.6
filetype==1.2.0
fireworks-ai==0.15.11
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-aws==0.2.10
langchain-core==0.3.29
langchain-fireworks==0.2.6
langchain-google-genai==2.0.8
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
monotonic==1.6
multidict==6.1.0
numpy==1.26.4
openai==1.59.6
orjson==3.10.14
packaging==24.2
pillow==11.1.0
playwright==1.49.1
posthog==3.7.5
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wsproto==1.2.0
yarl==1.18.3
